WORST: worst-case generation arg for merge sort In this commit, I introduce an argument -- WORST -- in the interpreter `qtest` for generating worst case for merge sorts. Supporting commands only required for `ih` and `it` in `qtest` interpreter. To be note that this work requires more experiments, which I didn't done much, to check if it's right.
To generate worst case for merge sort in `qtest`, it would initially
construct a list of continuous strings that start from string
`aaaaaaaaaa` by function `fill_cont_string()`. For example, if there is
an empty list generated by command `new`, after the command
`ih WORST 500` is called, `qtest` will first build a queue linked-list
[aaaaaaaatf aaaaaaaate aaaaaaaatd aaaaaaaatc ...]. Note that the maximum
list length depends on the number of distinct character changes in the
string length constant MAX_RANDSTR_LEN.

The decision to generate a string starting from `aaaaaaaaaa`, filling
MAX_RANDSTR_LEN, is based on reference [2]. According to [2], the worst
case for comparing two strings is to detect which is "bigger" only on
the last character. Therefore, I initialized the string starting from
`aaaaaaaaaa` and iteratively changed characters from the rightmost to
the leftmost while generating the linked list, aiming to fulfill the
requirements.

After the continious string is constructed, It will be reorganized to
the worst case of merge sort by function `worst_merge_split()`. The
thought of the code in `worst_merge_split()` refers to reference [1].
In [1], it indicates that the worst case of merge sort is to do the
maximum number of comparisons. The example work demonstrates splitting
the origin list to two lists with odd and even nodes from the origin
list, and recursively called the split, until there's no other nodes
that could be generate to two list with more than two nodes each.

In my work, I implement the function `worst_merge_split()` using a
bottom-up recursive approach. Similar to common merge sort
implementation, it will recursively called itself for the splitting
until there's no node in argument `head`, and the pointer `head->next`.
The construction of the splitting list is managed by pointer-to-pointer,
as well as the list joining method. There must exist a cache-friendly
implementation, but I currently can't thought of one.

Despite of the introduction of WORST arg, I also add two targets in
Makefile, which are called `sort_exp` and `sort_worst_exp`. The purpose
of these two targets is to automatically test the `q_sort` function in
random or worst-case scenarios, respectively, using perf, with the test
files located in the `test` and `test_worst` directories. On top of
that, I delete some needless comments in `queue.c`.

Ref:
 [1] When will the worst case of Merge Sort occur?
 https://stackoverflow.com/questions/24594112/when-will-the-worst-case-
 of-merge-sort-occur

 [2] Merge sort worst case running time for lexicographic
 sorting?
 https://stackoverflow.com/questions/9276163/merge-sort-worst-case-running
 -time-for-lexicographic-sorting

 [3] glibc/string/strcmp.c
 https://github.com/zerovm/glibc/blob/master/string/strcmp.c