`shuffle`: the Fisher-Yates shuffle algorithm According to the homework requirements, I implemented the command `shuffle` with Fisher-Yates shuffle algorithm[1], which is popularized by Donald E. Knuth in The Art of Computer Programming, to the interpretor `qtest`.
During the implementation, with the need of swapping nodes in
kernel-like linked-list, I introduced the `list_swap()` function from
Linux Kernel List Management API to the interpretor `qtest`. To minimize
the number of functions in the interpretor `qtest`, I replace the
function `list_replace()` in the original code to the actual approach it
was done in Linux Kernel List Management API.

For checking whether the implementation I done met the null hypothesis
test in chi-squared distribution or not, I also introduced the python
script in the homework description[2] as `shuffle_test.py` in this
commit.

Refs:
 [1] Fisher–Yates shuffle
 https://en.wikipedia.org/wiki/Fisher–Yates_shuffle

 [2] Homework description -- M01: lab0(D)
 https://hackmd.io/@sysprog/linux2024-lab0/%2F%40sysprog%2Flinux2024-lab0-d