`list_sort.[ch]`: The `lib/list_sort` file
For the sake of keeping the code in `queue.c` neat and clean, I decided
to move the `list_sort` that I imported from `lib/list_sort` out of
`queue.c`, and so did the future introduction for tim sort to this
repository.

Because of this change, I introduced a new comment `ksort`, representing
the `list_sort` from Linux kernel, in the interpreter `qtest`. The way
of the command `ksort` worked as same as that of `sort` command in
`qtest`. On top of that, I updated the `OBJ` macro in Makefile to
make it be able to compile `listsort.c`.