`timsort`: introduce the `minrun` strategy
For trying to enhance the stability and efficiency, and also to reduce
the comparisons of the original Timsort in `timsort.c`, I introduced
the definition of minimum run size, the `minrun` variable, into the
original code, to handle the minimum size of each runs. The definition
of the size of `minrun` follows the following discription in author's
manual[1]:

 - take the first `log2(MAX_MINRUN)` bits of `N`, and add 1 if any of
   the remaining bits are set. In fact, that rule covers every case in
   this section, including small `N` and exact powers of 2.

In my implementation, I set the MAX_MINRUN to be 32. Therefore, it would
take the first 5 bit of the input queue size.

Following the guidance of the author of Timsort, if a run is smaller
than the minrun variable, a technique fills elements until it reaches
the minimum run size. Then, a binary insertion sort is applied to sort
the elements. However, due to challenges in implementing binary
searching with linked-list data structures, I adopted an alternative
approach by inserting nodes using the "galloping" search technique
outlined in the author's manual[1] where it discribed how the
"galloping" works. Also, to be able to run the insertion sort properly,
I added a for loop to rebuild the `prev` pointer in every nodes,
preventing from the hidden infinite loop issue while doing the
insertion.

Refs:
 [1] cpython/Objects/listsort.txt
 https://github.com/python/cpython/blob/main/Objects/listsort.txt#L211