`sort_test`: The sorting test program for lab0-c
This program, `sort_test.[ch]`, conducts sorting tests for the node
structure in queue.h. It currently supports testing of the original Tim
sort code `timsort_old()` from course quiz 1, the Tim sort with linear
insertion sort min run implementation `timsort()`, `lib/listsort.c`, and
the sort `sort` in `queue.c`. Filenames are generated based on the
sorting algorithm, enabling multiple sorting tests with a single
command call in the interpreter qtest.c.

The goal of this test program is to determine whether the given sorting
algorithm is:
 - Stable: By checking if the seq variable within duplicate values
   serves in proper order.
 - Correctly sorting: By linearly traversing the list to confirm it is
   in ascending order.
Additionally, it records the number of comparisons and duration times
for the given test cases.

The test case generator in `sort_test.[ch]` closely resembles that in
the interpreter `qtest.c`, with a slight difference in the
implementation of duplicate case generation. After list generation,
if the list length is less than the constant 25000, a random shuffle
using Fisher-Yates algorithm is applied to the linked list, making it
more random. The choice of 25000 nodes as the threshold is to prevent
excessively long activation times.

To streamline `sort_test.[ch]`, some implementations for remaking the
output list to specific test case forms were moved to the program file
`sort_test_impl.[ch]`. This file contains code to reorganize the list
into worst-case form and implements the Fisher-Yates shuffle algorithm
for random shuffling in specific cases.

Generating worst-case scenarios is an interesting topic, especially in
character string comparison. Differences in string generation
significantly impact sorting duration time, even with the same number of
comparisons under the same generation algorithm. I've temporarily halted
worst-case generator testing until a method to measure the number of
comparisons during each sort is available. Further experimental results
will be committed as soon as possible, following discussions on
worst-case scenarios for merge sort in the course notes[1].

Reference:
 [1] 2024-03-{19,21} discussion note (In Chinese)
 https://hackmd.io/3e8LI8_0SLOo_UTDAnPKYQ?view#164253